:recycle: #130 ESLintの警告が出ている箇所を可能な限り修正
