+++ src/app/_components/ErrorTemplate/ErrorResetButton.tsx
@@ -7,6 +7,7 @@ export function ErrorResetButton({ resetFunc }: Props): JSX.Element {
     <button
       className="rounded-md bg-orange-500 px-3.5 py-2.5 text-sm font-semibold text-white shadow-sm hover:bg-orange-400 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-indigo-600"
       onClick={resetFunc}
+      type="button"
     >
       もう一度試す
     </button>
