+++ src/app/privacy/page.tsx
@@ -23,13 +23,11 @@ const PrivacyPage: NextPage = async () => {
   );
 
   return (
-    <>
-      <MarkdownContentsLayout>
-        <Header enableLoginLink={false} />
-        <MarkdownContents markdown={markdown} />
-        <Footer />
-      </MarkdownContentsLayout>
-    </>
+    <MarkdownContentsLayout>
+      <Header enableLoginLink={false} />
+      <MarkdownContents markdown={markdown} />
+      <Footer />
+    </MarkdownContentsLayout>
   );
 };
